Almost done with HLSL preprocessing. Looking into GLSL preprocessing next
