+++ include/nbl/asset/utils/CHLSLCompiler.h
@@ -36,7 +36,7 @@ class NBL_API2 CHLSLCompiler final : public IShaderCompiler
 			IShader::E_CONTENT_TYPE getCodeContentType() const override { return IShader::E_CONTENT_TYPE::ECT_HLSL; };
 		};
 
-		core::smart_refctd_ptr<ICPUShader> compileToSPIRV(const char* code, const IShaderCompiler::SCompilerOptions& options) const override;
+		core::smart_refctd_ptr<ICPUShader> compileToSPIRV_impl(const std::string_view code, const IShaderCompiler::SCompilerOptions& options, std::vector<IShaderCompiler::CCache::SEntry::SPreprocessingDependency>& dependencies) const override;
 
 		template<typename... Args>
 		static core::smart_refctd_ptr<ICPUShader> createOverridenCopy(const ICPUShader* original, const char* fmt, Args... args)
