+++ include/nbl/asset/utils/IShaderCompiler.h
@@ -12,7 +12,6 @@
 
 #include "nbl/asset/ICPUShader.h"
 #include "nbl/asset/utils/ISPIRVOptimizer.h"
-#include "nbl/asset/utils/SPreprocessingDependency.h"
 
 #include "nbl/core/xxHash256.h"
 
@@ -34,7 +33,6 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 					system::path absolutePath = {};
 					std::string contents = {};
 					std::optional<std::array<uint64_t, 4>> hash = std::nullopt; // TODO: we're not yet using IFile::getPrecomputedHash(), so for builtins we can maybe use that in the future
-					std::optional<nbl::system::IFileBase::time_point_t> lastWriteTime = std::nullopt;
 
 					explicit inline operator bool() const {return !absolutePath.empty();}
 				};
@@ -185,6 +183,42 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 			struct SEntry
 			{
 
+				struct SPreprocessingDependency
+				{
+					// Perf note: hashing while preprocessor lexing is likely to be slower than just hashing the whole array like this 
+					inline SPreprocessingDependency(const system::path& _requestingSourceDir, const std::string_view& _identifier, const std::string_view& _contents) :
+						requestingSourceDir(_requestingSourceDir), identifier(_identifier), contents(_contents)
+					{
+						assert(!_contents.empty());
+						const auto reqDirStr = requestingSourceDir.make_preferred().string();
+						std::vector<char> hashable;
+						hashable.insert(hashable.end(), reqDirStr.data()[0], reqDirStr.data()[reqDirStr.size()]);
+						hashable.insert(hashable.end(), identifier.data()[0], identifier.data()[identifier.size()]);
+						hashable.insert(hashable.end(), _contents.data()[0], _contents.data()[_contents.size()]);
+						// Can't static cast here?
+						hash = nbl::core::XXHash_256((uint8_t*)(hashable.data()), hashable.size() * (sizeof(char) / sizeof(uint8_t)));
+					}
+
+					inline SPreprocessingDependency(SPreprocessingDependency&&) = default;
+					inline SPreprocessingDependency& operator=(SPreprocessingDependency&&) = default;
+
+					inline bool operator==(const SPreprocessingDependency& other) const
+					{
+						return hash == other.hash && identifier == identifier && contents == contents;
+					}
+
+					// path or identifier
+					system::path requestingSourceDir = "";
+					std::string identifier = "";
+					// file contents
+					std::string contents = "";
+					// hash of the contents
+					std::optional<std::array<uint64_t, 4>> hash = std::nullopt;
+					// If true, then `getIncludeStandard` was used to find, otherwise `getIncludeRelative`
+					bool standardInclude = false;
+					nbl::system::IFileBase::time_point_t lastWriteTime = {};
+				};
+
 				// The ordering is important here, the dependencies MUST be added to the array IN THE ORDER THE PREPROCESSOR INCLUDED THEM!
 				using dependency_container_t = core::smart_refctd_dynamic_array<const SPreprocessingDependency>;
 				template<typename Container>
@@ -294,7 +328,7 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 			}
 
 			// can move to .cpp and have it not inline
-			inline asset::ICPUShader* find(const SPreprocessingDependency& mainFile, CIncludeFinder* finder) const
+			inline asset::ICPUShader* find(const SEntry::SPreprocessingDependency& mainFile, CIncludeFinder* finder) const
 			{
 				auto foundRange = m_container.equal_range(mainFile);
 				for (auto found = foundRange.first; found != foundRange.second; found++)
@@ -504,8 +538,7 @@ class NBL_API2 IShaderCompiler : public core::IReferenceCounted
 	protected:
 		virtual void insertIntoStart(std::string& code, std::ostringstream&& ins) const = 0;
 
-		// `mainDep` passed just so it doesn't have to be recomputed
-		virtual core::smart_refctd_ptr<ICPUShader> compileToSPIRV_impl(const std::string_view code, const SCompilerOptions& options, SPreprocessingDependency&& mainDep) const = 0;
+		virtual core::smart_refctd_ptr<ICPUShader> compileToSPIRV_impl(const std::string_view code, const SCompilerOptions& options, std::vector<CCache::SEntry::SPreprocessingDependency>& dependencies) const = 0;
 
 		core::smart_refctd_ptr<system::ISystem> m_system;
